+++ CHANGELOG.md
@@ -2,7 +2,23 @@
 
 All notable changes to `bolt` will be documented in this file
 
-## 0.1.0 - 2022-1-7
+## 0.1.1 - 2022-01-10
+
+- added the package `doctrine/dbal` for modding the database
+
+### fixes:
+
+- set default value if the categories are disabled
+- disable category_id validation if categories are disabled
+
+### improvements:
+
+#### `LetterResource`:
+
+- disable adding new Letter
+- set default replay message
+
+## 0.1.0 - 2022-01-07
 
 using [filament](https://filamentadmin.com/) as an admin panel, which give us: